Add check in removeDirectory for dir existence

+++ source/FileUtilities.c
@@ -161,11 +161,16 @@ boolByte removeDirectory(const CharString absolutePath) {
 #endif
 
 #if UNIX
-  // TODO: This is the lazy way of doing this...
+  if(!fileExists(absolutePath->data)) {
+    return false;
+  }
+
+  // This is a bit lazy, perhaps...
   CharString removeCommand = newCharString();
   snprintf(removeCommand->data, removeCommand->length, "/bin/rm -rf \"%s\"",
     absolutePath->data);
-  result = system(removeCommand->data) != 0;
+  result = system(removeCommand->data);
+  return (result == 0);
 #elif WINDOWS
   SHFILEOPSTRUCTA fileOperation = {0};
   fileOperation.wFunc = FO_DELETE;
@@ -176,8 +181,6 @@ boolByte removeDirectory(const CharString absolutePath) {
   logUnsupportedFeature("Copy directory recursively");
   return false;
 #endif
-
-  return result;
 }
 
 void buildAbsolutePath(const CharString directory, const CharString file, const char* fileExtension, CharString outString) {
